move File.Exists and File.ReadAllBytes to Common: remove code duplication and implementation differences
